unpack: Changed error to warning if filetype is not recognized. upack: Added copyfirst option.

+++ R/functions.r
@@ -389,32 +389,45 @@ abc <- function(x, capitals = FALSE){
 #' @param x character sting. file to be unpacked
 #' @param keeptar logical. If TRUE only decompresses the .tar.gz archive, otherwise fully unpack it?
 #' @param exdir character string. The directory to extract files to.
+#' @param copyfirst logical. Should the tar.gz file first be copied to the exdir directory?
 #' @param ... arguments passed from other method
 #' @return The default of unzip/untar/gunzip will be returned
 #' @description This funtcion tries to unpack a tar.gz or zip file
 #' @details If keeptar == TRUE, the .tar.gz archive is only beeing decompressed using the \code{\link{gunzip}}-function from the R.utils-package (to the .tar-file). Note that remove is set to FALSE, unlike the default of gunzip.
 #'
 #'    If keeptar == FALSE, the archive is fully unpacked using the \code{\link{untar}} function.
+#'
+#'    If copyfirst == TRUE (the default), the tar.gz file is copied to the exdir directory befor it is unpacked. Might be faster when unpacking from a slow source, e.g. an external harddrive. Only affects extraction of tar.gz files if kepptar == TRUE.
 #' @author Simon Frey
 #' @import R.utils
 #' @export
 #' @seealso \code{\link{unzip}}, \code{\link{untar}}, \code{\link{gunzip}}
-unpack <- function (x, keeptar = TRUE, exdir = ".", ...){
+unpack <- function (x, keeptar = TRUE, exdir = ".", copyfirst = TRUE, ...){
   extention <- unlist(strsplit(x, "[.]"))
   if (tail(extention, n = 1) == "gz") {
     if (tail(extention, n = 2)[1] != "tar") {
-      stop("Filetype not recognized")
-    }
-    else {
+      warning("Filetype not recognized")
+    } else {
       if (keeptar) {
         library(R.utils)
-        tarfile <- gunzip(as.character(x), remove = FALSE, ...)
-        if (exdir != ".") {
-          file.copy(from = tarfile, to = paste(TigeR::addSlash(exdir),
+        if(copyfirst){
+          if(exdir != "."){
+            copy <- paste(TigeR::addSlash(exdir), tail(unlist(strsplit(x, "/")), 1), sep = "")
+            file.copy(from = x, to = copy)
+            gunzip(as.character(copy), remove = TRUE, ...)
+          } else {
+            gunzip(as.character(x), remove = FALSE, ...)
+          }
+        } else {
+          tarfile <- gunzip(as.character(x), remove = FALSE, ...)
+          if (exdir != ".") {
+            file.copy(from = tarfile, to = paste(TigeR::addSlash(exdir),
                                                tail(unlist(strsplit(tarfile, "/")), 1),
                                                sep = ""))
-          file.remove(tarfile)
+            file.remove(tarfile)
+          }
         }
+
       }
       else {
         untar(x, exdir = exdir, ...)
@@ -425,7 +438,7 @@ unpack <- function (x, keeptar = TRUE, exdir = ".", ...){
     unzip(x, exdir = exdir, ...)
   }
   else {
-    stop("Filetype not recognized")
+    warning("Filetype not recognized")
   }
 }
